.pipeline/projects/youtube_workflow_tool/workspace/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "youtube-workflow-tool"
version = "0.1.0"
description = "YouTube video metadata generator, optimizer, and content planning tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
youtube-workflow = "youtube_workflow_tool.cli:cli"

[tool.setuptools.packages.find]
include = ["youtube_workflow_tool*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
